refactor(runner): split start idle lifecycle
